refactor and split input fn
Input has been decoupled into input_mouse and input_kb

Moved `lib.rs` around, ordered as enums > structs > plugin > functions
